# Conflicts:
#	src/main/java/com/codegym/quizappbackendmodule6/repository/QuizRepository.java
#	src/main/java/com/codegym/quizappbackendmodule6/repository/ResultRepository.java
#	src/main/java/com/codegym/quizappbackendmodule6/service/Impl/QuizServiceImpl.java
#	src/main/java/com/codegym/quizappbackendmodule6/service/QuizService.java
#	src/main/resources/application.properties

+++ src/main/java/com/codegym/quizappbackendmodule6/model/dto/QuizHistoryDTO.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -16,6 +17,6 @@ public class QuizHistoryDTO {
     private String quizName;
     private LocalDateTime finishTime;
     private long durationMinutes;
-    private Long score;
+    private BigDecimal score;
     private Integer attemptNumber;
 }
